trials/amazon-review-scraper/requirements.txt
python-dateutil
requests
selectorlib